Move PaginatedSequence to tools
