+++ item-impl/src/main/java/com/example/auction/item/impl/ItemServiceImpl.java
@@ -11,6 +11,7 @@ import com.example.auction.bidding.api.BiddingService;
 import com.example.auction.item.api.*;
 import com.example.auction.item.api.Item;
 import com.example.auction.item.impl.PItemCommand.*;
+import com.example.auction.pagination.PaginatedSequence;
 import com.lightbend.lagom.javadsl.api.ServiceCall;
 import com.lightbend.lagom.javadsl.api.broker.Topic;
 import com.lightbend.lagom.javadsl.api.deser.ExceptionMessage;
