+++ item-impl/src/test/java/com/example/auction/item/impl/ItemEventProcessorTest.java
@@ -3,7 +3,7 @@ package com.example.auction.item.impl;
 import com.datastax.driver.core.utils.UUIDs;
 import com.example.auction.item.api.ItemStatus;
 import com.example.auction.item.api.ItemSummary;
-import com.example.auction.item.api.PaginatedSequence;
+import com.example.auction.pagination.PaginatedSequence;
 import com.example.auction.item.impl.testkit.Await;
 import com.example.auction.item.impl.testkit.DoNothingTopicFactory;
 import com.example.auction.item.impl.testkit.ReadSideTestDriver;
