+++ web-gateway/app/views/paginate.scala.html
@@ -1,4 +1,4 @@
-@import com.example.auction.item.api.PaginatedSequence
+@import com.example.auction.pagination.PaginatedSequence
 
 @(pagination: PaginatedSequence[_])(route: (Int, String) => Html )
