+++ web-gateway/app/views/searchItem.scala.html
@@ -1,6 +1,7 @@
 @import helper._
-@import com.example.auction.item.api.{ItemStatus, PaginatedSequence}
+@import com.example.auction.item.api.ItemStatus
 @import com.example.auction.search.api.SearchItem
+@import com.example.auction.pagination.PaginatedSequence
 @import routes.ItemController.getItem
 @import java.util.Optional
 @import controllers.Currency
